+++ docsrc/releasenotes.rst
@@ -1,13 +1,17 @@
 Release Notes
 =============
 
-Version 1.0.3 (2025-12-02):
+Version 1.1 (2026-05-27):
 ++++++++++++++++++++++++++++
 - Added `AmplifierLinearityAnalysis` class for characterizing amplifier non-linearity.
+- Added `T1InversionRecovery` sequence.
+- Added rise time to linear chirp pulses.
 - Added right based arithmatic.
 - Fixed Version Detection Bug
 - Fixed Dependency issues
 - Improved Documentation
+- Removed Numba dependency
+
 
 
 Version 1.0.0 (2025-09-12):